Aggregate 0.13upgrade errors instead of exiting immediately
We have some repos that have Terraform code with intentional errors (e.g., a  syntax  error used for some automated test), which was problematic with the `tf-13upgrade.sh` script, as it would exit on the first such error. This PR upgrades the script to aggregate information about all errors, but keep running, and only show that information at the end.